requirements.txt
matplotlib==3.10.7
numpy==2.3.5
Pillow==12.0.0
streamlit==1.51.0
tensorflow==2.20.0
